Update and fixing some bugs and typo in laravel migration

+++ database/migrations/2025_07_06_181402_create_debt_histories_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('debt_histories', function (Blueprint $table) {
             $table->id();
-            $table->double('debt_nominal');
+            $table->decimal('debt_nominal',15,2);
             $table->dateTime('bill_date');
             $table->dateTime('due_date');
             $table->enum('status', ['pending', 'paid', 'late']);
